add Total Portfolio panel
