+++ examples-standalone/grid-live-data/src/app/app.component.html
@@ -1,35 +1,12 @@
 <app-header></app-header>
-<div >
-<my-portfolio></my-portfolio>
-<kendo-grid [kendoGridBinding]="(gridData | async)!" [rowClass]="rowCallback">
-    <kendo-grid-column field="id" title="#" [width]="35"> </kendo-grid-column>
-    <kendo-grid-column field="currency" title="Crypto currency">
-        <ng-template kendoGridCellTemplate let-dataItem>
-            <div class="currency-container">
-                <img class="coin-logo" src="../assets/coinslogo/{{ dataItem.currency }}.png" alt="coin logo" />
-                <span class="currency">
-                    {{ dataItem.currency }}
-                </span>
-                <span class="crypto-symbol"> ({{ dataItem.symbol }}) </span>
-            </div>
-        </ng-template>
-    </kendo-grid-column>
-    <kendo-grid-column field="currentPrice" title="Price">
-        <ng-template kendoGridCellTemplate let-dataItem>
-                <span [@positiveState]="dataItem.currentPrice > prevDataItem.currentPrice ? true : false" [@negativeState]="dataItem.currentPrice < prevDataItem.currentPrice ? true : false">
-                    {{ dataItem.currentPrice | currency}}
-                </span>
-        </ng-template>
-         </kendo-grid-column>
-    <kendo-grid-column class="priceColumn" field="change_24h" title="24h%">
-        <ng-template kendoGridCellTemplate let-dataItem>
-            <span> {{ dataItem.change_24h | number }}% </span>
-            <span class="k-icon k-i-arrow-60-down arrow-down" *ngIf="dataItem.change_24h < prevDataItem.change_24h"></span>
-            <span class="k-icon k-i-arrow-60-up arrow-up" *ngIf="dataItem.change_24h > prevDataItem.change_24h"></span>
-        </ng-template>
-    </kendo-grid-column>
+<div style="display: flex">
+    <div>
+        <my-portfolio></my-portfolio>
+        <total-portfolio></total-portfolio>
+        <grid></grid>
+    </div>
 
-    <kendo-grid-column field="volume" title="Market Cap" format="c"> </kendo-grid-column>
-</kendo-grid>
-<app-main-panel></app-main-panel>
+    <div>
+        <app-main-panel></app-main-panel>
+    </div>
 </div>
