+++ examples-standalone/grid-live-data/src/app/app.module.ts
@@ -8,21 +8,25 @@ import { NavigationModule } from '@progress/kendo-angular-navigation';
 import { IndicatorsModule } from '@progress/kendo-angular-indicators';
 import { IconsModule } from '@progress/kendo-angular-icons';
 import { LayoutModule } from '@progress/kendo-angular-layout';
+import { ChartsModule } from '@progress/kendo-angular-charts';
+import { ButtonsModule } from '@progress/kendo-angular-buttons';
+import 'hammerjs';
 
 import { AppComponent } from './app.component';
-import { HeaderComponent } from './header/header/header.component';
-
+import { HeaderComponent } from './header/header.component';
+import { TotalPortfolioComponent } from './total-portfolio/total-portfolio.component';
 import { MyPortfolioComponent } from './my-portfolio/my-portfolio.component';
-import { MainPanelComponent } from './main-panel/main-panel/main-panel.component';
-import { ButtonsModule } from '@progress/kendo-angular-buttons';
-
+import { MainPanelComponent } from './main-panel/main-panel.component';
+import { GridComponent } from './grid/grid.component';
 
 @NgModule({
   declarations: [
     AppComponent,
     HeaderComponent,
     MyPortfolioComponent,
-    MainPanelComponent
+    MainPanelComponent,
+    GridComponent,
+    TotalPortfolioComponent
   ],
   imports: [
     BrowserModule,
@@ -34,6 +38,7 @@ import { ButtonsModule } from '@progress/kendo-angular-buttons';
     IconsModule,
     LayoutModule,
     ButtonsModule,
+    ChartsModule,
   ],
   providers: [],
   bootstrap: [AppComponent]
