+++ examples-standalone/grid-live-data/src/app/grid/grid.component.css
@@ -1,0 +1,27 @@
+.price-down .priceColumn,
+.arrow-down {
+    color: #f31700;
+}
+
+.price-up .priceColumn,
+.arrow-up {
+    color: #37b400;
+}
+
+.coin-logo {
+    height: 22px;
+    width: 22px;
+}
+.currency {
+    margin: 3px 5px 0px 5px;
+    color: #4b5ffa;
+    font-weight: bold;
+}
+
+.currency-container {
+    display: flex;
+    align-items: center;
+}
+.crypto-symbol {
+    color: #959595;
+}
