+++ examples-standalone/grid-live-data/src/app/header/header.component.ts
