+++ examples-standalone/grid-live-data/src/app/main-panel/transaction-data/transactions.ts
