+++ examples-standalone/grid-live-data/src/app/services/stocks.service.ts
@@ -10,9 +10,6 @@ export class StocksService {
     private immutableData!: Stock[];
     public previousData: Stock[] = [];
 
-    private cardsUrl: string = 'assets/cardsData.json';
-    private cardsData!: Stock[];
-
     public updateFreq: number = 2000;
 
     constructor(private http: HttpClient) {}
@@ -32,20 +29,6 @@ export class StocksService {
             });
         });
     }
-
-    getCards(): Observable<Stock[]> {
-      return new Observable<Stock[]>((observer) => {
-          this.http.get<Stock[]>(this.cardsUrl).subscribe((data: Stock[]) => {
-              this.cardsData = data;
-              observer.next(this.cardsData);
-              setInterval(() => {
-                this.cardsData = this.cardsData.map((row: Stock) => this.updateRandomRowWithData(row));
-
-                  observer.next(this.cardsData);
-              }, this.updateFreq);
-          });
-      });
-  }
 
     updateRandomRowWithData(row: Stock): Stock {
         const shouldUpdateData = Math.random() < 0.3;
