+++ examples-standalone/grid-live-data/src/app/total-portfolio/total-portfolio.component.ts
@@ -1,0 +1,9 @@
+import { Component, ViewEncapsulation } from '@angular/core';
+
+@Component({
+    selector: 'total-portfolio',
+    templateUrl: './total-portfolio.component.html',
+    styleUrls: ['./total-portfolio.component.css'],
+    encapsulation: ViewEncapsulation.None
+})
+export class TotalPortfolioComponent {}
